tools/headers: Fix undefined behaviour (34 << 26)
Left-shifting past the size of your datatype is undefined behaviour in
C.  The literal 34 gets the type `int`, and that one is not big enough
to be left shifted by 26 bits.

An `unsigned` is long enough (on any machine that has at least 32 bits
for their ints.)

For uniformity, we mark all the literals as unsigned.  But it's only
really needed for HUGETLB_FLAG_ENCODE_16GB.

Thanks to Randy Dunlap for an initial review and suggestion.

Signed-off-by: Matthias Goergens <matthias.goergens@gmail.com>